ollama_discord/requirements.txt
requests
python-dotenv
discord.py
pyperclip
psutil
pyautogui
Pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
